src/client/requirements.txt
outside-framework==1.5.6
